Work on modules.
It's far from perfect yet, but we're leveraging Go here for modules. The
basic idea is that you'll reference modules as Go packages, we will
compile them for the host OS+arch, and upload the static go binary.

We're only doing that because Go allows you to cross-compile trivially
with Go 1.5+, and it lets you make a static executable. Go totally
totally sucks as a language for writing this stuff, though.